requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
streamlit>=1.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
langgraph>=0.2.0
python-dotenv>=1.0.0
